backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
django-storages==1.14.2
azure-storage-blob==12.19.0
azure-cognitiveservices-speech==1.34.0
openai==1.3.0
requests==2.31.0
PyJWT==2.8.0
Pillow==10.1.0
python-magic==0.4.27
reportlab==4.0.7
pydub==0.25.1
gunicorn==21.2.0
